+++ CHANGELOG.md
@@ -2,4 +2,11 @@
 
 ## 0.1.0.0 -- 2022-07-26
 
-* First version.
+### Added
+
+* `validatorSucceedsWith`, ensuring validator runs successfully given datum, redeemer, and script context.
+* `validatorFailsWith`, ensuring validator fails given datum, redeemer, and script context.
+* `mintingPolicySucceedsWith`, ensuring minting policy runs successfully given datum, redeemer, and script context.
+* `mintingPolicyFailsWith`, ensuring minting policy fails given datum, redeemer, and script context.
+* `scriptSucceeds`, ensuring `Script` runs successfully.
+* `scriptFails`, ensuring `Script` fails.